Quick startmodule i18 (#25)
* Add missing provider to addons

* Add the ability to specify a workspace path so a separate repo can be used for workspace configurations

* Add missing kube2iam config file to addons

* Remove default workspaces which are provider specific.
Add examples quickstart with tf setup and example workspace which can be
used in a deployers own repository to configure the environment

* Update documentation, and makefile variable names
consistent with new workspaces setup

* Change make target to create to be consistent with terraform vocab - create/destroy

* Update deployment instructions to match make create target

* Install external-dns
WIP: There some configuration variables in the external-dns.yml that are deployer specific.
Need to discuss domain and certificate handling

* Update Makefile

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Update Makefile

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Update Makefile

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Update Makefile

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Update Makefile

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Change create to apply in makefile and scripts to be fully consistent with terraform terminology
Add  make clean-terraform target to delete local terraform directories on request rather than all the time

* Move from zone terminology to domain name and hostname
Remove domain and certificate creation (formally zones) as this is now handled by external-dns and ingress
Change external_dns configuration to support variables for domain name and txtownerid
Note: This required removing the external-dns.yaml file and embedding it in the external_dns.tf file. Terraform Helm Provider was throwing errors about passing in the value using set  with all the combinations of escape sequences and string manipulation trialled. It's ugly this way but it works.

* Replace availability zone list with aws data provider

* Update addons/_variables.tf

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Update addons/external_dns.tf

Co-Authored-By: tom-butler <tombutler.development@gmail.com>

* Update apply.sh

Co-Authored-By: roarmstrong <34013278+roarmstrong@users.noreply.github.com>

* Update destroy.sh

Co-Authored-By: roarmstrong <34013278+roarmstrong@users.noreply.github.com>

* Put back the pod annotations so external-dns can work

* Remove default terraform state bucket name to ensure deployer takes responsibility for unqiuely naming the bucket