Terraform aws odc install kubectl (#152)
* Add a null-resource to install kubectl into workstations that don't have it installed e.g. TF Cloud
* update installation of CRD to use local state so future destroys are all good

IMPORTANT:
Applying this update to your running infrastructure is a two step process

If you simply try to apply it your existing state will not have the local vars used by the Destroy provider. More importantly, whatever your previous Destroy provider did won't be undone when the existing resource is replaced.
Two options:
1. Before you update your TF code, destroy the k8s infrastructure (easiest but drastic)
2. Comment out the destroy provider and replace it with the one you are currently using -> apply this, this will update the local state vars ready for the new destroy provider -> put in the new destroy provider

This occurs because the execution code for the destroy provider is read from the TF source, but the state vars are read from the current state. So you get a chicken egg problem if you simply add the new destroy provider as the null-resource will be replaced to add the new state vars on first run, and they don't exist yet.