Terraform aws odc v1.8 (#179)
Release v1.8 datacube-k8s-eks

This is release v1.8 of the datacube-k8s-eks terraform aws infrastructure as code.
This is a MAJOR refactor and is not suitable for in-place update of an existing deployment using previous versions
Fortunately you can do a test deployment right alongside your existing one so long as you take care to set your configuration to avoid collisions in domain names, cluster name, etc. I won't explain how, if you don't know how to do this you shouldn't be operating terraform ;-)

**Negative effects of this change**
Will making this change break or change an existing functionality? flag it here

This is a new release with a new approach. Indeed this repo is now a terraform module repository with an example of a live repo in it. You can look at that to see how you would start with your own live repo using this module repo.

Why 1.8?
because its a first official release and starting numbers are fairly arbitrary so we decided to be consistent with ODC datacube-core 1.8 release numbering.

Read the docs!
Co-authored-by: Nikita Gandhi <nikita.r.gandhi@gmail.com>
Co-authored-by: Tisham Dhar <whatnickd@gmail.com>
Co-authored-by: tom-butler <tombutler.development@gmail.com>
Co-authored-by: whatnick <tisham@whatnick.com>